requirements.txt
altair
dateparser
cherrypy
fitbit
inflection
matplotlib
numpy
pandas
prettytable
plaid-python
pyaml
oauthlib==2.1.0
requests-oauthlib==1.1.0
seaborn
splitwise
stravalib
stravaio
streamlit
watchdog
